refactor: 대체 된 기존 테스트 삭제
